Reconfigured how the jokes appear on the DOM
